% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_refresh.R
\name{get_info_covid19swiss}
\alias{get_info_covid19swiss}
\title{Get information about the datasets provided by covid19swiss}
\usage{
get_info_covid19swiss()
}
\value{
a data.frame of information about the datasets in this package
}
\description{
Returns information about the datasets in this package for
  covid19R harvesting
}
\examples{
\donttest{

get_info_covid19swiss()
}

}
