% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_refresh.R
\name{refresh_covid19swiss}
\alias{refresh_covid19swiss}
\title{Refresh the 2019 Novel Coronavirus COVID-19 (2019-nCoV) Dataset in the Covid19R Project Format}
\source{
Specialist Unit for Open Government Data Canton of Zurich \href{https://www.zh.ch/de/politik-staat/opendata.html}{website}
}
\usage{
refresh_covid19swiss()
}
\value{
A data.frame object
}
\description{
aily summary of the Coronavirus (COVID-19) pandemic cases in Switzerland by Canton
}
\examples{
\donttest{
# update the data
covid19_swiss <- refresh_covid19swiss()
}

}
